requirements.txt
feedparser
newspaper3k
sentence-transformers
psycopg2-binary
scikit-learn